encoder: added retry mechanism on vdi_init [1/1]
PD#SWPL-12344

Problem:
Encoder may open fail when open twice simultaniously.
it is due to driver allocate memory and set to busy at that time.

Solution:
Test APP have the wait and retry mechanism, it is clearer to put
it on user library also to avoid the integration issues.

Verify:
AE400

Change-Id: I6e72924ce2c66f73fd6893e63806d8f1d85e952a
Signed-off-by: yongchun.li <yongchun.li@amlogic.com>